Proyecto_3_Grupo_5/airflow/requirements.txt
pandas
mlflow
psycopg2-binary
boto3

